feature(client): advanced search refactor part1
