+++ client/actions/events.js
@@ -4,7 +4,7 @@ import moment from 'moment-timezone'
 import * as selectors from '../selectors'
 import { SubmissionError } from 'redux-form'
 
-const receiveEvents = (events) => ({
+export const receiveEvents = (events) => ({
     type: 'RECEIVE_EVENTS',
     payload: events,
     receivedAt: Date.now()
@@ -135,71 +135,10 @@ function _fetchEvents({ keyword, ids }) {
 
 export function fetchEvents({ keyword }) {
     return (dispatch, getState, { $timeout, $location }) => {
-        dispatch({ type: 'REQUEST_EVENTS' })
+        dispatch({ type: 'REQUEST_EVENTS', payload: { name: keyword } })
         dispatch(_fetchEvents({ keyword }))
         .then(data => dispatch(receiveEvents(data._items)))
         // update the url (deep linking)
         .then(() => $timeout(() => ($location.search('searchEvent', keyword))))
-    }
-}
-
-function _searchEvents({ form }) {
-    return (dispatch, getState, { api }) => {
-        const query = {}
-        const range = {}
-        const filter = {}
-        const should = []
-
-        if (form.name) {
-            should.push(
-                { match: { name: form.name } },
-                { match: { definition_short: form.name } }
-            )
-        }
-
-        if (form.location) {
-            should.push(
-                { match: { 'location.name': form.location } },
-                { match: { 'location.qcode': form.location } }
-            )
-        }
-
-        if (form.dates) {
-            if (form.dates.start) {
-                range['dates.start'] = { gte: form.dates.start }
-            }
-
-            if (form.dates.end) {
-                range['dates.end'] = { lte: form.dates.end }
-            }
-
-            filter.range = range
-        }
-
-        // build the query
-        if (should.length > 0) {
-            query.bool = { should: should }
-        }
-
-        if (filter.length > 0) {
-            query.filter = filter
-        }
-
-        // Query the API and sort by date
-        return api('events').query({
-            sort: '[("dates.start",1)]',
-            embedded: { files: 1 },
-            source: JSON.stringify({ query, filter })
-        })
-    }
-}
-
-export function searchEvents({ form }) {
-    return (dispatch, getState, { $timeout, $location }) => {
-        dispatch({ type: 'REQUEST_EVENTS' })
-        dispatch(_searchEvents({ form }))
-        .then(data => dispatch(receiveEvents(data._items)))
-        // update the url (deep linking)
-        .then(() => $timeout(() => ($location.search('advancedSearchEvent', form))))
     }
 }
