+++ client/actions/index.js
@@ -1,4 +1,5 @@
 export * from './events'
+export * from './search'
 export * from './locations'
 export * from './modal'
 export * from './planning'
