+++ client/containers/AdvancedSearchPanelContainer/style.scss
@@ -1,5 +1,5 @@
 
-$ADVANCED_SEARCH_WIDTH: 450px;
+$ADVANCED_SEARCH_WIDTH: 580px;
 
 .Planning__advanced-search {
     position: relative;
