+++ client/containers/EventsListPanelContainer/style.scss
@@ -1,3 +1,17 @@
+@import '~superdesk-core/styles/sass/variables.scss';
+
+.Planning__events-list-container {
+    height: 100%;
+}
+
 .Planning__events-list {
+    position: relative;
     display: flex;
+    transition-property: width;
+    transition-duration: .25s;
+    padding-top: $nav-height;
+    box-sizing: border-box;
+    border-right: 3px solid #666;
+    box-shadow: 2px 0 10px 0 rgba(0,0,0,.3);
+    height: 100%;
 }
