+++ client/containers/PlanningApp.jsx
@@ -6,11 +6,9 @@ import {
     AddEventContainer,
     CreateAgendaContainer } from './index'
 
-const PlanningAppComponent = ({ editPlanningViewOpen, advancedSearchViewOpen, showEvents }) => {
+const PlanningAppComponent = ({ showEvents }) => {
     const classes = [
         'Planning',
-        editPlanningViewOpen ? 'Planning--edit-planning-view' : null,
-        advancedSearchViewOpen ? 'Planning--advanced-search-view' : null,
         showEvents ? 'Planning--show-events' : null,
     ]
     return (
@@ -24,15 +22,11 @@ const PlanningAppComponent = ({ editPlanningViewOpen, advancedSearchViewOpen, sh
 }
 
 PlanningAppComponent.propTypes = {
-    editPlanningViewOpen: React.PropTypes.bool,
-    advancedSearchViewOpen: React.PropTypes.bool,
-    showEvents: React.PropTypes.bool,
+    showEvents: React.PropTypes.bool
 }
 
 const mapStateToProps = (state) => ({
     showEvents: state.events.show,
-    editPlanningViewOpen: state.planning.editorOpened,
-    advancedSearchViewOpen: state.planning.advancedSearchOpened
 })
 
 export const PlanningApp = connect(mapStateToProps)(PlanningAppComponent)
