+++ client/containers/PlanningPanelContainer/style.scss
@@ -1,27 +1,31 @@
 @import '~superdesk-core/styles/sass/variables.scss';
 
-.Planning__planning {
-    display: flex;
-    .subnav {
-        .subnav__page-title { width: 100px; }
-        select {
-            top: 10px;
-            position: relative;
+.Planning--edit-planning-container {
+    width: 100%;
+    .Planning__planning {
+        display: flex;
+        height: 100%;
+        .subnav {
+            .subnav__page-title { width: 100px; }
+            select {
+                top: 10px;
+                position: relative;
+            }
         }
-    }
-    &__list {
-        flex-grow: 1;
-        .list-view { box-shadow: inherit; }
-    }
-    .ListItem__list-item {
-        width: auto; // otherwise it overflow on the right because of the margin
-        margin: 10px 20px;
-    }
-    &__empty-message {
-        font-size: 1.75em;
-        line-height: 1.2em;
-        text-align: center;
-        margin-top: 40px;
+        &__list {
+            flex-grow: 1;
+            .list-view { box-shadow: inherit; }
+        }
+        .ListItem__list-item {
+            width: auto; // otherwise it overflow on the right because of the margin
+            margin: 10px 20px;
+        }
+        &__empty-message {
+            font-size: 1.75em;
+            line-height: 1.2em;
+            text-align: center;
+            margin-top: 40px;
+        }
     }
     .Planning__edit-planning {
         position: relative;
